requirements.txt
pandas
openpyxl
pandasql
python-dotenv
openai
